refactor: rename new feature and cleanup
