+++ crates/node-api/tests/endpoints.rs
@@ -1,6 +1,6 @@
 use essential_node::{self as node};
 use essential_node_api as node_api;
-use essential_node_types::BlockTx;
+use essential_node_types::block_notify::BlockTx;
 use essential_types::{convert::bytes_from_word, Block, Value};
 use futures::{StreamExt, TryStreamExt};
 use tokio_util::{
