+++ crates/node-types/src/lib.rs
@@ -12,6 +12,10 @@ use essential_types::{
 };
 use serde::{Deserialize, Serialize};
 
+/// Wrappers around tokio's `watch` channel for notifying of new blocks.
+#[cfg(feature = "block-notify")]
+pub mod block_notify;
+
 /// The default big-bang configuration.
 pub const DEFAULT_BIG_BANG: &str = include_str!("../big-bang.yml");
 
@@ -67,21 +71,6 @@ pub struct BigBang {
     pub solution: Solution,
 }
 
-#[cfg(feature = "new-block")]
-/// Wrapper around `watch::Sender` to notify of new blocks.
-///
-/// This is used by `essential-builder` to notify `essential-relayer`
-/// and by `essential-relayer` to notify [`validation`] stream.
-#[derive(Clone, Default)]
-pub struct BlockTx(tokio::sync::watch::Sender<()>);
-
-#[cfg(feature = "new-block")]
-/// Wrapper around `watch::Receiver` to listen to new blocks.
-///
-/// This is used by [`db::subscribe_blocks`] stream.
-#[derive(Clone)]
-pub struct BlockRx(tokio::sync::watch::Receiver<()>);
-
 impl BigBang {
     /// Produce the big bang [`Block`].
     pub fn block(&self) -> Block {
@@ -97,40 +86,6 @@ impl Default for BigBang {
     fn default() -> Self {
         serde_yaml::from_str(DEFAULT_BIG_BANG)
             .expect("default `big-bang-block.yml` must be valid (checked in tests)")
-    }
-}
-
-#[cfg(feature = "new-block")]
-impl BlockTx {
-    /// Create a new [`BlockTx`] to notify listeners of new blocks.
-    pub fn new() -> Self {
-        let (block_tx, _block_rx) = tokio::sync::watch::channel(());
-        Self(block_tx)
-    }
-
-    /// Notify listeners that a new block has been received.
-    ///
-    /// Note this is best effort and will still send even if there are currently no listeners.
-    pub fn notify(&self) {
-        let _ = self.0.send(());
-    }
-
-    /// Create a new [`BlockRx`] to listen for new blocks.
-    pub fn new_listener(&self) -> BlockRx {
-        BlockRx(self.0.subscribe())
-    }
-
-    /// Get the number of receivers listening for new blocks.
-    pub fn receiver_count(&self) -> usize {
-        self.0.receiver_count()
-    }
-}
-
-#[cfg(feature = "new-block")]
-impl BlockRx {
-    /// Waits for a change notification.
-    pub async fn changed(&mut self) -> Result<(), tokio::sync::watch::error::RecvError> {
-        self.0.changed().await
     }
 }
